+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Released: -
 * [New] `Pico::prepareFileContent()` and `Pico::substituteFileContent()` both
         now receive the (optional) `$pageId` argument for the new `%page_*%`
         Markdown placeholders
+* [New] Add `page()` Twig function to access a page's data
 * [Changed] ! Pico now requires PHP 7.2.5 or later (this includes full PHP 8
             support, also see #528, #534, #608)
 * [Changed] ! Pico now depends on Twig 3.3, skipping Twig 2.x altogether; this